+++ lib/teiserver/coordinator/consul_server.ex
@@ -913,17 +913,11 @@ defmodule Teiserver.Coordinator.ConsulServer do
     match_id = Battle.get_lobby_match_id(state.lobby_id)
 
     block_status = Account.check_block_status(userid, member_list)
-
-    # Temporary code while it is tested
-    case block_status do
-      :blocked ->
-        Logger.error("User is blocked")
-      :blocking ->
-        Logger.error("User is blocking")
-      _ ->
-        :ok
-    end
-    block_status = :ok
+    boss_avoid_status = state.host_bosses
+      |> Stream.map(fn boss_id ->
+        Account.does_a_avoid_b?(boss_id, userid)
+      end)
+      |> Enum.any?
 
     cond do
       client == nil ->
@@ -963,6 +957,10 @@ defmodule Teiserver.Coordinator.ConsulServer do
       block_status == :blocked ->
         Telemetry.log_simple_lobby_event(userid, match_id, "join_refused.blocked")
         {false, "You are blocked by too many players in this lobby"}
+
+      boss_avoid_status == true ->
+        Telemetry.log_simple_lobby_event(userid, match_id, "join_refused.boss_blocked")
+        {false, "You are blocked by the boss of this lobby"}
 
       true ->
         {true, nil}